Simplify S3OddMod2.

+++ Decompose.Numerics/Core/UInt128.cs
@@ -193,22 +193,22 @@ namespace Decompose.Numerics
             return c;
         }
 
-        public static UInt128 operator &(UInt128 a, uint b)
+        public static uint operator &(UInt128 a, uint b)
         {
             return a.r0 & b;
         }
 
-        public static UInt128 operator &(uint a, UInt128 b)
+        public static uint operator &(uint a, UInt128 b)
         {
             return a & b.r0;
         }
 
-        public static UInt128 operator &(UInt128 a, ulong b)
+        public static ulong operator &(UInt128 a, ulong b)
         {
             return a.s0 & b;
         }
 
-        public static UInt128 operator &(ulong a, UInt128 b)
+        public static ulong operator &(ulong a, UInt128 b)
         {
             return a & b.r0;
         }
